# Node-Backend/swagger.json
{
	"components": {
...
		"schemas": {
			"ColorInformation": {
				"properties": {
					"color": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"color"
				],
				"type": "object"
			},
			"ModeInformation": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Rainbow"
					},
					"mode": {
						"type": "integer",
						"format": "int32",
						"example": 11,
						"minimum": 0
					}
				},
				"required": [
					"mode"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "strype_backend",
		"version": "1.0.0",
		"description": "Node backend for Strype",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Stein Jonker"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/color": {
			"post": {
				"operationId": "SetColor",
				"responses": {
					"202": {
						"description": "Accepted"
					}
				},
				"tags": [
					"Color"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ColorInformation"
							}
						}
					}
				}
			}
		},
		"/mode": {
			"post": {
				"operationId": "SetMode",
				"responses": {
					"202": {
						"description": "Accepted"
					}
				},
				"tags": [
					"Mode"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ModeInformation"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetModes",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ModeInformation"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Mode"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}